authenticação e autorização e inicio dos testes e swagger

+++ vr-card/src/main/java/br/com/vr/models/Card.java
@@ -2,12 +2,12 @@ package br.com.vr.models;
 
 import lombok.Builder;
 import lombok.Getter;
-import lombok.NoArgsConstructor;
 import lombok.Setter;
 import org.springframework.data.annotation.Id;
 import org.springframework.data.mongodb.core.index.Indexed;
 import org.springframework.data.mongodb.core.mapping.Document;
 
+import java.math.BigDecimal;
 import java.time.Instant;
 
 @Document(collection = "cards")
@@ -21,6 +21,6 @@ public class Card {
     private String numberCard;
     private String HashPassword;
     private String userId;
-    private Double cardBalance;
+    private BigDecimal cardBalance;
     private Instant createdAt;
 }
